# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Product API"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
      "/allProducts": {
        "get": {
          "summary": "Retrieve all products",
          "description": "Returns a list of all products",
          "responses": {
            "200": {
              "description": "A list of products",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Product"
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/allproducts/{id}": {
        "get": {
          "summary": "Get product by ID",
          "description": "Returns a product with the specified ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the product to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Product found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            }
          }


        }
        },
      "/getmessages": {
        "get": {
          "summary": "Retrieve all Messages",
          "description": "Returns a list of allMessages ",
          "responses": {
            "200": {
              "description": "A list of messages"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/getmessages/{id}": {
        "delete": {
          "summary": "Delete a message by its ID",
          "description": "Deletes a message by its ID from the system",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the message to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Message deleted successfully"
            },
            "404": {
              "description": "Message not found"
            }
          }
        }
      },"/users": {
        "get": {
          "summary": "Retrieve all Users",
          "description": "Returns a list of all users",
          "responses": {
            "200": {
              "description": "A list of products",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/users/{id1}": {
        "get": {
          "summary": "Get user by ID",
          "description": "Returns a user with the specified ID",
          "parameters": [
            {
              "name": "id1",
              "in": "path",
              "description": "ID of the product to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "user found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            },
            "404": {
              "description": "user not found"
            }
          }
        }
    },
    "/users/{id}": {
      "delete": {
        "summary": "Delete a user by its ID",
        "description": "Deletes a user by its ID from the system",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the user to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "user found"
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
    "/register":{
        "post": {
          "summary": "Registers a new user",
          "tags": [
            "Authentication"
          ],
          "description": "Registers a new user with the provided information.",
          "parameters": [
            {
              "name": "username",
              "in": "formData",
              "description": "Name of the user",
              "required": true,
              "type": "string"
            },
            {
              "name": "email",
              "in": "formData",
              "description": "Email address of the user",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "description": "Password of the user",
              "required": true,
              "type": "string",
              "format": "password"
            }
          ],
          "responses": {
            "200": {
              "description": "User registration successful"
            },
            "400": {
              "description": "Bad request"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
},
    "definitions": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Product ID"
          },
          "title": {
            "type": "string",
            "description": "Product name"
          },
          "desc": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "image": {
            "type": "string",
            "description": "Product image"
          }
        }
      }
    },"User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "description": "Username"
        },
        "email": {
          "type": "string",
          "description": "Email"
        },
        "password": {
          "type": "string",
          "description": "Password"
        }
      }
    }
  }